+++ docs/changelog.md
@@ -5,6 +5,7 @@
 * Implement network-controller backend
 * Implement support for sending and receiving MAC commands (no support for proprietary commands yet)
 * Refactor test scenarios
+* Web-interface: nodes can now be accessed from the applications tab (nodes button)
 
 **Note:** You need to update to LoRa Semtech Bridge 2.0.1+ or 1.1.4+ since
 it fixes a mac command related marshaling issue.